Revert "refactoring detect"
This reverts commit d1dc825f2e6dc37a97a1d872963dfae69e3836ec.
